+++ test/product-sitemap-pipe.test.js
@@ -25,17 +25,6 @@ const DEFAULT_CONFIG = {
   owner: 'adobe',
   repo: 'helix-pages',
   ref: 'main',
-  public: {
-    patterns: {
-      base: {
-        storeViewCode: 'default',
-        storeCode: 'main',
-      },
-      '/products/{{urlKey}}': {
-        pageType: 'product',
-      },
-    },
-  },
 };
 
 const DEFAULT_STATE = (opts = {}) => (new PipelineState({
@@ -95,7 +84,7 @@ describe('Product Sitemap Pipe Test', () => {
 
   it('handles a 404', async () => {
     const s3Loader = new FileS3Loader();
-    s3Loader.rewrite('default.json', 'missing-file-404.json');
+    s3Loader.rewrite('index.json', 'missing-file-404.json');
 
     const state = DEFAULT_STATE({
       s3Loader,
@@ -108,7 +97,7 @@ describe('Product Sitemap Pipe Test', () => {
       new PipelineRequest(new URL('https://acme.com/products2/sitemap.xml')),
     );
     assert.strictEqual(result.status, 404);
-    assert.strictEqual(result.headers.get('x-error'), 'failed to load /products/sitemap.xml from product-bus: 404');
+    assert.strictEqual(result.headers.get('x-error'), 'failed to load org/site/indices/products/index.json from product-bus: 404');
   });
 
   it('returns 404 for invalid path info', async () => {
@@ -198,8 +187,6 @@ describe('Product Sitemap Pipe Test', () => {
               },
             },
             route: {
-              storeCode: 'main',
-              storeViewCode: 'default',
               matchedPatterns: ['/products/{{urlKey}}'],
             },
           },
@@ -235,8 +222,6 @@ describe('Product Sitemap Pipe Test', () => {
               },
             },
             route: {
-              storeCode: 'main',
-              storeViewCode: 'default',
               matchedPatterns: ['/products/{{urlKey}}'],
             },
           },
@@ -272,25 +257,11 @@ describe('Product Sitemap Pipe Test', () => {
 </urlset>`);
     });
 
-    it('resolves location from config if no url in product data', () => {
+    it('omits products without url or path', () => {
       const xml = toSitemapXML(
         {
           prodHost: 'https://www.example.com',
-          config: {
-            public: {
-              patterns: {
-                '/products/{{urlKey}}': {
-                  storeCode: 'main',
-                  storeViewCode: 'default',
-                },
-              },
-            },
-            route: {
-              storeCode: 'main',
-              storeViewCode: 'default',
-              matchedPatterns: ['/products/{{urlKey}}'],
-            },
-          },
+          config: {},
         },
         {
           foo: {
@@ -305,13 +276,42 @@ describe('Product Sitemap Pipe Test', () => {
       );
       assert.deepStrictEqual(xml, `<?xml version="1.0" encoding="utf-8"?>
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:xhtml="http://www.w3.org/1999/xhtml">
+</urlset>`);
+    });
+
+    it('constructs location from path when url is not provided', () => {
+      const xml = toSitemapXML(
+        {
+          prodHost: 'https://www.example.com',
+          config: {
+            public: {
+              productSitemapConfig: {
+                lastmod: 'YYYY-MM-DD',
+              },
+            },
+          },
+        },
+        {
+          foo: {
+            data: {
+              id: 'foo',
+              description: 'This is a description',
+              path: '/products/foo-product',
+              lastModified: '2021-04-30T03:47:18Z',
+            },
+          },
+        },
+      );
+      assert.deepStrictEqual(xml, `<?xml version="1.0" encoding="utf-8"?>
+<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:xhtml="http://www.w3.org/1999/xhtml">
   <url>
-    <loc>https://www.example.com/products/foo-url-key</loc>
+    <loc>https://www.example.com/products/foo-product</loc>
+    <lastmod>2021-04-30</lastmod>
   </url>
 </urlset>`);
     });
 
-    it('allows setting an extension on urls, if url is not provided in product data', () => {
+    it('constructs location from path with extension when url is not provided', () => {
       const xml = toSitemapXML(
         {
           prodHost: 'https://www.example.com',
@@ -319,12 +319,8 @@ describe('Product Sitemap Pipe Test', () => {
             public: {
               productSitemapConfig: {
                 extension: '.html',
+                lastmod: 'YYYY-MM-DD',
               },
-            },
-            route: {
-              storeCode: 'main',
-              storeViewCode: 'default',
-              matchedPatterns: ['/products/{{urlKey}}'],
             },
           },
         },
@@ -333,7 +329,7 @@ describe('Product Sitemap Pipe Test', () => {
             data: {
               id: 'foo',
               description: 'This is a description',
-              urlKey: 'foo-url-key',
+              path: '/products/foo-product',
               lastModified: '2021-04-30T03:47:18Z',
             },
           },
@@ -342,20 +338,21 @@ describe('Product Sitemap Pipe Test', () => {
       assert.deepStrictEqual(xml, `<?xml version="1.0" encoding="utf-8"?>
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:xhtml="http://www.w3.org/1999/xhtml">
   <url>
-    <loc>https://www.example.com/products/foo-url-key.html</loc>
+    <loc>https://www.example.com/products/foo-product.html</loc>
+    <lastmod>2021-04-30</lastmod>
   </url>
 </urlset>`);
     });
 
-    it('omits products with no matching pattern and no url', () => {
+    it('adds https protocol to prodHost if not present', () => {
       const xml = toSitemapXML(
         {
-          prodHost: 'https://www.example.com',
+          prodHost: 'www.example.com',
           config: {
-            route: {
-              storeCode: 'main',
-              storeViewCode: 'default',
-              matchedPatterns: null, // should never actually happen
+            public: {
+              productSitemapConfig: {
+                lastmod: 'YYYY-MM-DD',
+              },
             },
           },
         },
@@ -364,7 +361,7 @@ describe('Product Sitemap Pipe Test', () => {
             data: {
               id: 'foo',
               description: 'This is a description',
-              urlKey: 'foo-url-key',
+              path: '/products/foo-product',
               lastModified: '2021-04-30T03:47:18Z',
             },
           },
@@ -372,6 +369,10 @@ describe('Product Sitemap Pipe Test', () => {
       );
       assert.deepStrictEqual(xml, `<?xml version="1.0" encoding="utf-8"?>
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:xhtml="http://www.w3.org/1999/xhtml">
+  <url>
+    <loc>https://www.example.com/products/foo-product</loc>
+    <lastmod>2021-04-30</lastmod>
+  </url>
 </urlset>`);
     });
   });